actualizacion de cuestionario

+++ database/migrations/2020_08_14_022654_create_rel_trabajador_datos_laborales_nomina.php
@@ -16,7 +16,7 @@ class CreateRelTrabajadorDatosLaboralesNomina extends Migration
         Schema::create('rel_trabajador_datos_laborales_nomina', function (Blueprint $table) {
             $table->smallIncrements('id')->unsigned();
             $table->smallInteger('trabajador_id')->unsigned();
-            $table->smallInteger('institucion_id')->default(1)->unsigned()->nullable();
+            $table->mediumInteger('institucion_id')->default(1)->unsigned()->nullable();
             $table->string('codigo_puesto_id')->nullable();
             $table->mediumInteger('codigo_puesto_sinerhias')->nullable();
             $table->string('descripcion_puesto', 254)->nullable();
@@ -25,7 +25,7 @@ class CreateRelTrabajadorDatosLaboralesNomina extends Migration
             $table->mediumInteger('entidad_federativa_puesto_id')->default(7)->unsigned()->nullable();
             $table->mediumInteger('tipo_contrato_id')->unsigned()->nullable();
             $table->mediumInteger('tipo_plaza_id')->unsigned()->nullable();
-            $table->smallInteger('institucion_puesto_id')->default(1)->unsigned()->nullable();
+            $table->mediumInteger('institucion_puesto_id')->default(1)->unsigned()->nullable();
             $table->string('ur', 100)->nullable();
             $table->mediumInteger('unidad_administradora_id')->default(2032563)->unsigned()->nullable();
             $table->smallInteger('tipo_nomina_id')->unsigned()->nullable();
